rework to fake a new http.Request and a http.ResponseWriter * Add support for context propagation * Normalize span tag use

Apply comments

Add "t.Helper" to helper methods
