+++ contrib/envoyproxy/go-control-plane/envoy_test.go
@@ -370,11 +370,12 @@ func newEnvoyAppsecRig(t *testing.T, traceClient bool, interceptorOpts ...ddgrpc
 
 	interceptorOpts = append([]ddgrpc.InterceptorOption{ddgrpc.WithServiceName("grpc")}, interceptorOpts...)
 
-	server := grpc.NewServer()
+	server := grpc.NewServer(
+		grpc.StreamInterceptor(StreamServerInterceptor(interceptorOpts...)),
+	)
 
 	fixtureServer := new(envoyFixtureServer)
-	appsecSrv := AppsecEnvoyExternalProcessorServer(fixtureServer)
-	envoyextproc.RegisterExternalProcessorServer(server, appsecSrv)
+	envoyextproc.RegisterExternalProcessorServer(server, fixtureServer)
 
 	li, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
